ingrediant add, changed database
