+++ app/src/main/java/team3/recipefinder/database/AppDatabase.kt
@@ -4,9 +4,12 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
 import team3.recipefinder.dao.CookbookDao
 import team3.recipefinder.dao.RecipeDao
 import team3.recipefinder.model.*
+
 
 @Database(entities = [
     Recipe::class,
@@ -23,5 +26,5 @@ abstract class AppDatabase: RoomDatabase() {
 }
 
  fun getAppDatabase(context: Context): AppDatabase {
-    return Room.databaseBuilder(context, AppDatabase::class.java, "app_database").build()
+    return Room.databaseBuilder(context, AppDatabase::class.java, "app_database1").build()
 }
