+++ app/src/main/java/team3/recipefinder/viewmodel/RecipeDetailViewModel.kt
@@ -62,7 +62,7 @@ class RecipeDetailViewModel(
 
     private suspend fun addI(i: Long) {
         withContext(Dispatchers.IO) {
-            database.assignIngredientToRecipe(i, recipe.value!!.id)
+            database.assignIngredientToRecipe(i, recipe.value!!.id,"aa")
         }
     }
 
@@ -77,6 +77,18 @@ class RecipeDetailViewModel(
         uiScope.launch {
             val step = RecipeStep(0, name)
             addS(step)
+        }
+    }
+    fun addIngredient(name: String) {
+        uiScope.launch {
+            val recipe = Ingredient(0, name)
+            addI(recipe)
+        }
+    }
+
+    private suspend fun addI(i: Ingredient) {
+        withContext(Dispatchers.IO) {
+            database.insertIngredient(i)
         }
     }
