+++ app/src/main/java/team3/recipefinder/viewmodel/RecipeViewModel.kt
@@ -26,21 +26,8 @@ class RecipeViewModel(val database: RecipeDao, application: Application) :
 
     fun addRecipe(name: String) {
         uiScope.launch {
-            val recipe = Recipe(0, name)
+            val recipe = Recipe(0, name,"Test Description")
             addR(recipe)
-        }
-    }
-
-    fun addIngredient(name: String) {
-        uiScope.launch {
-            val recipe = Ingredient(0, name)
-            addI(recipe)
-        }
-    }
-
-    private suspend fun addI(i: Ingredient) {
-        withContext(Dispatchers.IO) {
-            database.insertIngredient(i)
         }
     }
