+++ app/src/main/res/layout/dialog_set_list.xml
@@ -22,13 +22,37 @@
         app:layout_constraintGuide_end="16dp" />
 
 
+    <Button
+        android:id="@+id/button"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:onClick="showAddRecipeDialog"
+        android:tag="@string/text_ingredientName"
+        android:text="addIngrediant"
+        tools:layout_editor_absoluteX="32dp"
+        tools:layout_editor_absoluteY="0dp" />
+
+    <EditText
+        android:id="@+id/editText"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:ems="10"
+        android:inputType="textPersonName"
+        android:text="Name"
+        app:layout_constraintTop_toBottomOf="@+id/radiogroup"
+        tools:layout_editor_absoluteX="16dp" />
+
     <RadioGroup
         android:id="@+id/radiogroup"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_centerHorizontal="true"
         android:layout_centerVertical="true"
-        android:orientation="vertical" />
+        android:orientation="vertical"
+        app:layout_constraintTop_toBottomOf="@+id/button"
+        tools:layout_editor_absoluteX="0dp">
+
+    </RadioGroup>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
